Changed some stuff
Leaving the speed entry empty now sets the speed to the default value
instead of freezing the board.

Replaced the x and y arguments to the get_neighbors method with a single
argument

Replaced the count_neighbors method with a one liner using conjunction
of sets.

Removed some empty lines

Fixed/improved a lot of comments
added a docstring

Renamed dead_previously to potentially_born

This is pretty much a test if I can remember what I changed.